Separate types and functions
