+++ src/core/types.rs
@@ -1,0 +1,11 @@
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize)]
+pub struct User {
+    id: i64,
+    is_bot: bool,
+    first_name: String,
+    last_name: Option<String>,
+    username: Option<String>,
+    language_code: Option<String>,
+}
